==> code/00-Genomic-Location-of-DML.ipynb <==
# Genomic Location of DML

In this notebook, I will identify the genomic locations of [sex-specific DML identified with `methylKit`](https://github.com/epigeneticstoocean/2018_L18-adult-methylation/blob/main/code/03.4-methylkit.Rmd). 

## 0. Set working directory

pwd

cd ../output/

mkdir DML-characterization

cd DML-characterization/

bedtoolsDirectory = "/opt/homebrew/bin/"

## 1. Remove C->T SNPs from DML

I will count how many DML overlap with SNPs, then remove those overlapping DML before proceeding with analyses.

### 1a. Create BEDfile

!awk '{print $1"\t"$2"\t"$2}' ../../genome-features/unique-CT-SNPs.tab \
> ../../genome-features/unique-CT-SNPs.bed

!head ../../genome-features/unique-CT-SNPs.bed
!wc -l ../../genome-features/unique-CT-SNPs.bed

### 1b. Identify overlaps

#Remove SNPs from female DML list and save as a new file
!{bedtoolsDirectory}subtractBed \
-a ../../data/female_dml.bed \
-b ../../genome-features/unique-CT-SNPs.bed \
> ../../data/female_dml-NO-SNPs.bed
!head ../../data/female_dml-NO-SNPs.bed

#Compare # DML with and without SNPs
!wc -l ../../data/female_dml.bed
!wc -l ../../data/female_dml-NO-SNPs.bed

#Remove SNPs from male DML list and save as a new file
!{bedtoolsDirectory}subtractBed \
-a ../../data/male_dml.bed \
-b ../../genome-features/unique-CT-SNPs.bed \
> ../../data/male_dml-NO-SNPs.bed
!head ../../data/male_dml-NO-SNPs.bed

#Compare # DML with and without SNPs
!wc -l ../../data/male_dml.bed
!wc -l ../../data/male_dml-NO-SNPs.bed

femaleDML = "../../data/female_dml-NO-SNPs.bed"

maleDML = "../../data/male_dml-NO-SNPs.bed"

#Count the number of female hypomethylated DML
#Count the number of female hypermethylated DML
!grep "-" {femaleDML} | wc -l
!grep -v "-" {femaleDML} | wc -l

#Count the number of male hypomethylated DML
#Count the number of male hypermethylated DML
!grep "-" {maleDML} | wc -l
!grep -v "-" {maleDML} | wc -l

#Look at common DML between female and male lists
!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b {maleDML} \

## 2. Characterize genomic locations of DML

I will look at overlaps between genome features and either female- or male-DML.

### 3a. Gene

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-gene.gff \
> female_dml-Gene.bed
!head female_dml-Gene.bed
!wc -l female_dml-Gene.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-gene.gff \
> male_dml-Gene.bed
!head male_dml-Gene.bed
!wc -l male_dml-Gene.bed

I also want to get DML overlaps with genes that include gene ID or transcript ID. This will be useful for downstream annotation.

!head ../../genome-features/C_virginica-3.0-gene.gff

#Find overlaps between DML and gene list with annotations
#Save overlaps with annotations

!{bedtoolsDirectory}intersectBed \
-wb \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-gene.gff \
> female_dml-Gene-wb.bed
!head female_dml-Gene-wb.bed

#Find overlaps between DML and gene list with annotations
#Save overlaps with annotations

!{bedtoolsDirectory}intersectBed \
-wb \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-gene.gff \
> male_dml-Gene-wb.bed
!head male_dml-Gene-wb.bed

### 3b. Exon UTR

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-exonUTR.gff \
> female_dml-exonUTR.bed
!head female_dml-exonUTR.bed
!wc -l female_dml-exonUTR.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-exonUTR.gff \
> male_dml-exonUTR.bed
!head male_dml-exonUTR.bed
!wc -l male_dml-exonUTR.bed

### 3c. CDS

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-CDS.gff \
> female_dml-CDS.bed
!head female_dml-CDS.bed
!wc -l female_dml-CDS.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-CDS.gff \
> male_dml-CDS.bed
!head male_dml-CDS.bed
!wc -l male_dml-CDS.bed

### 3d. Intron

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-intron.bed \
> female_dml-intron.bed
!head female_dml-intron.bed
!wc -l female_dml-intron.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-intron.bed \
> male_dml-intron.bed
!head male_dml-intron.bed
!wc -l male_dml-intron.bed

### 3e. Upstream flanks

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-upstream.gff \
> female_dml-upstream.bed
!head female_dml-upstream.bed
!wc -l female_dml-upstream.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-upstream.gff \
> male_dml-upstream.bed
!head male_dml-upstream.bed
!wc -l male_dml-upstream.bed

### 3f. Downstream flanks

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-downstream.gff \
> female_dml-downstream.bed
!head female_dml-downstream.bed
!wc -l female_dml-downstream.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-downstream.gff \
> male_dml-downstream.bed
!head male_dml-downstream.bed
!wc -l male_dml-downstream.bed

### 3g. Intergenic regions

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-intergenic.bed \
> female_dml-intergenic.bed
!head female_dml-intergenic.bed
!wc -l female_dml-intergenic.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-intergenic.bed \
> male_dml-intergenic.bed
!head male_dml-intergenic.bed
!wc -l male_dml-intergenic.bed

### 3h. lncRNA

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-lncRNA.gff \
> female_dml-lncRNA.bed
!head female_dml-lncRNA.bed
!wc -l female_dml-lncRNA.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-lncRNA.gff \
> male_dml-lncRNA.bed
!head male_dml-lncRNA.bed
!wc -l male_dml-lncRNA.bed

### 3i. Tranposable elements

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {femaleDML} \
-b ../../genome-features/C_virginica-3.0-rm.te.bed \
> female_dml-TE.bed
!head female_dml-TE.bed
!wc -l female_dml-TE.bed

#Find overlaps between DML and feature
#Look at output
#Count number of overlaps

!{bedtoolsDirectory}intersectBed \
-u \
-a {maleDML} \
-b ../../genome-features/C_virginica-3.0-rm.te.bed \
> male_dml-TE.bed
!head male_dml-TE.bed
!wc -l male_dml-TE.bed

## 4. Combine line counts

This will make it easier for downstream analysis.

!find female_dml*bed

#Get line count for all DML overlap files
#Remove the 10th line (total entries)
#Print in a tab-delimited format
#Save output

!wc -l female_dml*bed \
| sed '10,$ d' \
| awk '{print $1"\t"$2}' \
> female_dml-Overlap-counts.txt
!head female_dml-Overlap-counts.txt

!find male_dml*bed

#Get line count for all DML overlap files
#Remove the 10th line (total entries)
#Print in a tab-delimited format
#Save output

!wc -l male_dml*bed \
| sed '10,$ d' \
| awk '{print $1"\t"$2}' \
> male_dml-Overlap-counts.txt
!head male_dml-Overlap-counts.txt
