==> code/16-DML-annotation.Rmd <==
---
title: "16-DML-annotation"
output: html_document
---


Working with to dml files

```{bash}
head ../data/*dml.bed
```

```{bash}
cat ../data/male_dml.bed | awk -v OFS='\t' '{print $1, $2, $3}' > ../output/16-DML-annotation/male-dml-simple.bed
```







```{bash}
wc -l ../data/*dml.bed
```



```{bash}
cd ../genome-features 

curl -O  https://eagle.fish.washington.edu/Cvirg_tracks/C_virginica-3.0_Gnomon_genes.bed
```


Where do these DMLs overlap in genome? 

```{bash}
/home/shared/bedtools2/bin/intersectBed \
-wb \
-a ../data/female_dml.bed \
-b ../genome-features/C_virginica-3.0_Gnomon_genes.bed | wc -l
```


```{bash}
/home/shared/bedtools2/bin/intersectBed \
-wb \
-a ../data/female_dml.bed \
-b ../genome-features/C_virginica-3.0_Gnomon_genes.bed | head
```








```{bash}
/home/shared/bedtools2/bin/intersectBed \
-wb \
-a ../data/female_dml.bed \
-b ../genome-features/C_virginica-3.0_Gnomon_genes.bed | awk '{print $9}' | awk -F'-' '{print $2}' > ../output/16-DML-annotation/female_dml_geneID.txt
```


```{bash}
/home/shared/bedtools2/bin/intersectBed \
-wb \
-a ../data/male_dml.bed \
-b ../genome-features/C_virginica-3.0_Gnomon_genes.bed | awk '{print $9}' | awk -F'-' '{print $2}' > ../output/16-DML-annotation/male_dml_geneID.txt
```
