==> code/42-predominant_isoform-female_male.ipynb <==
## Wrangle [ceabigr](https://github.com/sr320/ceabigr) predominant isoform data in females and males

---

### List computer specs

%%bash
echo "TODAY'S DATE"
date
echo "------------"
echo ""
lsb_release -a
echo ""
echo "------------"
echo "HOSTNAME: "
hostname
echo ""
echo "------------"
echo "Computer Specs:"
echo ""
lscpu
echo ""
echo "------------"
echo ""
echo "Memory Specs"
echo ""
free -mh

## Set variables
%env line=--------------------------------------------------------------

## Identify differences in predominant isoforms between female and male sets

### Peek at the data sets we're working with

%%bash

head ../output/42-predominant-isoform/predom_iso-[FM]*.txt

echo ""
echo "${line}"
echo "Line counts for each file:"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-FEMALE.txt

wc -l ../output/42-predominant-isoform/predom_iso-MALE.txt

### Identify unique entries in the FEMALE and MALE predominant isoform files

Uses `comm` command to _suppress_ common lines between two files.

%%bash

# Prints unique lines present in the first file (FEMALE)
# Skips header line (NR > 1)
comm -23 \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-FEMALE.txt | sort) \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-MALE.txt | sort) \
> ../output/42-predominant-isoform/predom_iso-FEMALE-unique.txt

# Prints unique lines present in the second file (MALE)
comm -13 \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-FEMALE.txt | sort) \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-MALE.txt | sort) \
> ../output/42-predominant-isoform/predom_iso-MALE-unique.txt

# Peek at each file
head ../output/42-predominant-isoform/*unique.txt

echo ""
echo "${line}"
echo "Line counts for each file:"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-FEMALE-unique.txt

wc -l ../output/42-predominant-isoform/predom_iso-MALE-unique.txt

echo ""
echo "${line}"
echo ""

### Confirm no difference in genes identified between two files

No output means there are _no_ differences.

%%bash
diff \
<(awk '{print $1}' ../output/42-predominant-isoform/predom_iso-FEMALE-unique.txt) \
<(awk '{print $1}' ../output/42-predominant-isoform/predom_iso-MALE-unique.txt)

## Create file with column indicating different predominant isoform between sexes

0 - No difference

1 - Difference

Join on `predom_iso-FEMALE.txt` and keep all of the `predom_iso-FEMALE.txt` lines.

If a line has 3 fields, then there was _no_ match and gets a `0` in output file.

If a line has 4 fields, then there _was_ a match and gets a `1` in output file.

%%bash
# Join and keep all lines from first file (-a 1)
join \
-a 1 \
<(awk 'NR > 1' ../output/42-predominant-isoform/predom_iso-FEMALE.txt | sort -k1,1) \
<(sort -k1,1 ../output/42-predominant-isoform/predom_iso-FEMALE-unique.txt) \
| awk '{if (NF==4) print $1, "\t", 1; else print $1, "\t", 0}' \
> ../output/42-predominant-isoform/predom_iso-diff-female_vs_male.tab

# Check file format
head ../output/42-predominant-isoform/*diff-female_vs_male.tab



echo ""
echo "${line}"
echo "Confirm there are are lines with a "1" in second column."
echo ""
echo "Expecting 4090:"
echo ""

# Display counts of lines with 0 and 1 in second column
awk '{print $2}' ../output/42-predominant-isoform/*diff-female_vs_male.tab \
| sort \
| uniq --count

echo ""
echo "${line}"
echo "Line counts for predom_iso-diff-female_vs_male.tab (expecting 39014):"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-diff-female_vs_male.tab

## Identify differences in predominant isoforms between _control_ and _exposed_ FEMALE sets.

### Peek at the data sets we're working with

%%bash

head ../output/42-predominant-isoform/predom_iso-[ce]*FEMALE.txt

echo ""
echo "${line}"
echo "Line counts for each file:"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-controls-FEMALE.txt

wc -l ../output/42-predominant-isoform/predom_iso-exposed-FEMALE.txt

### Identify unique entries in the control FEMALE and exposed FEMALE predominant isoform files

Uses `comm` command to _suppress_ common lines between two files.

%%bash

# Prints unique lines present in the first file (controls FEMALE)
# Skips header line (NR > 1)
comm -23 \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-controls-FEMALE.txt | sort) \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-exposed-FEMALE.txt | sort) \
> ../output/42-predominant-isoform/predom_iso-controls-FEMALE-unique.txt

# Prints unique lines present in the second file (exposed FEMALE)
comm -13 \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-controls-FEMALE.txt | sort) \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-exposed-FEMALE.txt | sort) \
> ../output/42-predominant-isoform/predom_iso-exposed-FEMALE-unique.txt

# Peek at each file
head ../output/42-predominant-isoform/predom_iso-[ce]*FEMALE-unique.txt

echo ""
echo "${line}"
echo "Line counts for each file:"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-controls-FEMALE-unique.txt

wc -l ../output/42-predominant-isoform/predom_iso-exposed-FEMALE-unique.txt

echo ""
echo "${line}"
echo ""

### Confirm no difference in genes identified between two files

No output means there are _no_ differences.

%%bash
diff \
<(awk '{print $1}' ../output/42-predominant-isoform/predom_iso-controls-FEMALE-unique.txt) \
<(awk '{print $1}' ../output/42-predominant-isoform/predom_iso-exposed-FEMALE-unique.txt)

## Create file with column indicating different predominant isoform between controls/exposed FEMALE

0 - No difference

1 - Difference

Join on `predom_iso-control-FEMALE.txt` and keep all of the `predom_iso-control-FEMALE.txt` lines.

If a line has 3 fields, then there was _no_ match and gets a `0` in output file.

If a line has 4 fields, then there _was_ a match and gets a `1` in output file.

%%bash
# Join and keep all lines from first file (-a 1)
join \
-a 1 \
<(awk 'NR > 1' ../output/42-predominant-isoform/predom_iso-controls-FEMALE.txt | sort -k1,1) \
<(sort -k1,1 ../output/42-predominant-isoform/predom_iso-controls-FEMALE-unique.txt) \
| awk '{if (NF==4) print $1, "\t", 1; else print $1, "\t", 0}' \
> ../output/42-predominant-isoform/predom_iso-diff-controls_vs_exposed-female.tab

# Check file format
head ../output/42-predominant-isoform/*diff-controls_vs_exposed-female.tab



echo ""
echo "${line}"
echo "Confirm there are are lines with a "1" in second column."
echo ""
echo "Expecting 2252:"
echo ""

# Display counts of lines with 0 and 1 in second column
awk '{print $2}' ../output/42-predominant-isoform/*diff-controls_vs_exposed-female.tab \
| sort \
| uniq --count

echo ""
echo "${line}"
echo "Line counts for predom_iso-diff-female_vs_male.tab (expecting 39014):"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-diff-controls_vs_exposed-female.tab

## Identify differences in predominant isoforms between _control_ and _exposed_ MALE sets.

### Peek at the data sets we're working with

%%bash

head ../output/42-predominant-isoform/predom_iso-[ce]*-MALE.txt

echo ""
echo "${line}"
echo "Line counts for each file:"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-controls-MALE.txt

wc -l ../output/42-predominant-isoform/predom_iso-exposed-MALE.txt

### Identify unique entries in the control and exposed MALE predominant isoform files

Uses `comm` command to _suppress_ common lines between two files.

%%bash

# Prints unique lines present in the first file (controls MALE)
# Skips header line (NR > 1)
comm -23 \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-controls-MALE.txt | sort) \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-exposed-MALE.txt | sort) \
> ../output/42-predominant-isoform/predom_iso-controls-MALE-unique.txt

# Prints unique lines present in the second file (exposed MALE)
comm -13 \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-controls-MALE.txt | sort) \
<(awk 'NR > 1 {print $1, $2}' ../output/42-predominant-isoform/predom_iso-exposed-MALE.txt | sort) \
> ../output/42-predominant-isoform/predom_iso-exposed-MALE-unique.txt

# Peek at each file
head ../output/42-predominant-isoform/predom_iso-[ce]*-MALE-unique.txt

echo ""
echo "${line}"
echo "Line counts for each file:"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-controls-MALE-unique.txt

wc -l ../output/42-predominant-isoform/predom_iso-exposed-MALE-unique.txt

echo ""
echo "${line}"
echo ""

### Confirm no difference in genes identified between two files

No output means there are _no_ differences.

%%bash
diff \
<(awk '{print $1}' ../output/42-predominant-isoform/predom_iso-controls-MALE-unique.txt) \
<(awk '{print $1}' ../output/42-predominant-isoform/predom_iso-exposed-MALE-unique.txt)

## Create file with column indicating different predominant isoform between controls/exposed MALE

0 - No difference

1 - Difference

Join on `predom_iso-control-MALE.txt` and keep all of the `predom_iso-control-MALE.txt` lines.

If a line has 3 fields, then there was _no_ match and gets a `0` in output file.

If a line has 4 fields, then there _was_ a match and gets a `1` in output file.

%%bash
# Join and keep all lines from first file (-a 1)
join \
-a 1 \
<(awk 'NR > 1' ../output/42-predominant-isoform/predom_iso-controls-MALE.txt | sort -k1,1) \
<(sort -k1,1 ../output/42-predominant-isoform/predom_iso-controls-MALE-unique.txt) \
| awk '{if (NF==4) print $1, "\t", 1; else print $1, "\t", 0}' \
> ../output/42-predominant-isoform/predom_iso-diff-controls_vs_exposed-male.tab

# Check file format
head ../output/42-predominant-isoform/*diff-controls_vs_exposed-male.tab



echo ""
echo "${line}"
echo "Confirm there are are lines with a "1" in second column."
echo ""
echo "Expecting 1808:"
echo ""

# Display counts of lines with 0 and 1 in second column
awk '{print $2}' ../output/42-predominant-isoform/*diff-controls_vs_exposed-male.tab \
| sort \
| uniq --count

echo ""
echo "${line}"
echo "Line counts for predom_iso-diff-female_vs_male.tab (expecting 39014):"
echo ""

wc -l ../output/42-predominant-isoform/predom_iso-diff-controls_vs_exposed-male.tab
